cross-compile to scala 3

+++ azure/src/main/scala/blobstore/azure/AzureStore.scala
@@ -4,21 +4,21 @@ package azure
 import blobstore.url.{Path, Url}
 import cats.effect.{Async, Resource}
 import cats.effect.std.Queue
-import cats.syntax.all._
+import cats.syntax.all.*
 import com.azure.core.util.FluxUtil
 import com.azure.storage.blob.batch.BlobBatchClientBuilder
 import com.azure.storage.blob.{BlobContainerAsyncClient, BlobServiceAsyncClient}
-import com.azure.storage.blob.models.{BlobItemProperties, _}
+import com.azure.storage.blob.models.*
 import com.azure.storage.common.implementation.Constants
 import fs2.{Chunk, Pipe, Stream}
-import fs2.interop.reactivestreams._
+import fs2.interop.reactivestreams.*
 import reactor.core.publisher.{Flux, Mono}
 
 import java.nio.ByteBuffer
 import java.time.Duration
 import java.time.temporal.ChronoUnit
-import java.util.function.{Function => JavaFunction}
-import scala.jdk.CollectionConverters._
+import java.util.function.{Function as JavaFunction}
+import scala.jdk.CollectionConverters.*
 
 /** @param azure - Azure Blob Service Async Client
   * @param defaultFullMetadata – return full object metadata on [[list]], requires additional request per object.
@@ -79,7 +79,7 @@ class AzureStore[F[_]: Async](
         .getBlobContainerAsyncClient(container)
         .getBlobAsyncClient(blobName)
       val flux = Flux.from(publisher)
-      val pto  = new ParallelTransferOptions().setBlockSizeLong(blockSize).setMaxConcurrency(numBuffers.max(2))
+      val pto  = new ParallelTransferOptions().setBlockSizeLong(blockSize.toLong).setMaxConcurrency(numBuffers.max(2))
       val (overwriteCheck, requestConditions) =
         if (overwrite) {
           Mono.empty -> null // scalafix:ok
@@ -168,7 +168,7 @@ class AzureStore[F[_]: Async](
       blobClient = azure
         .getBlobContainerAsyncClient(container)
         .getBlobAsyncClient(blob)
-      pto    = new ParallelTransferOptions().setBlockSizeLong(blockSize).setMaxConcurrency(numBuffers.max(2))
+      pto    = new ParallelTransferOptions().setBlockSizeLong(blockSize.toLong).setMaxConcurrency(numBuffers.max(2))
       flux   = Flux.from(publisher)
       upload = blobClient.upload(flux, pto, true)
       _ <-
